SIGA_T/requirements.txt
validators==0.20.0
torch==1.7.0
torchvision==0.8.0
timm==0.4.12
lmdb==1.2.1
pillow==8.1.0
nltk==3.7
natsort==8.2.0
opencv-python==4.5.1.48
opencv-contrib-python==4.5.1.48
wand==0.6.7
transformers==4.2.1
strsimpy==0.2.1